+++ opendut-carl/src/peer/broker.rs
@@ -67,15 +67,6 @@ impl PeerMessagingBroker {
         Ok(())
     }
 
-    /// deprecated since version 0.2.0
-    pub async fn list_peers(&self) -> Vec<PeerId> {
-        let peers = self.peers.read().await;
-
-        peers.keys()
-            .cloned()
-            .collect::<Vec<_>>()
-    }
-
     pub async fn open(
         &self,
         peer_id: PeerId,