+++ tests/src/register.rs
@@ -1,7 +1,6 @@
 use anyhow::anyhow;
 use opendut_util::telemetry;
-use std::ops::Not;
-
+use opendut_types::peer::state::PeerState;
 use crate::util;
 
 #[tokio::test(flavor = "multi_thread")]
@@ -10,20 +9,19 @@ async fn register_edgar_carl() -> anyhow::Result<()> {
 
     let carl_port = util::spawn_carl()?;
 
-    util::spawn_edgar(carl_port)?;
+    let edgar_id = util::spawn_edgar(carl_port)?;
 
     let carl_client = util::spawn_carl_client(carl_port).await?;
 
     util::retry(|| async {
         let mut carl_client = carl_client.lock().await;
 
-        let peers = carl_client.broker.list_peers().await
+        let edgar_state = carl_client.peers.get_peer_state(edgar_id).await
             .map_err(|cause| backoff::Error::transient(cause.into()))?;
 
-        if peers.is_empty().not() {
-            Ok(())
-        } else {
-            Err(backoff::Error::transient(anyhow!("No peers registered in time!")))
+        match edgar_state {
+            PeerState::Up { .. } => Ok(()),
+            PeerState::Down => Err(backoff::Error::transient(anyhow!("No peers registered in time!")))
         }
     }).await?;